Extract contract_artifacts to separate mod
