+++ crates/extrinsics/src/lib.rs
@@ -16,6 +16,7 @@
 
 mod balance;
 mod call;
+mod contract_artifacts;
 mod contract_info;
 mod env_check;
 mod error;
@@ -30,16 +31,9 @@ mod upload;
 #[cfg(feature = "integration-tests")]
 mod integration_tests;
 
-use colored::Colorize;
 use env_check::compare_node_env_with_contract;
 
-use anyhow::{
-    anyhow,
-    Context,
-    Result,
-};
-use std::path::PathBuf;
-
+use anyhow::Result;
 use contract_build::{
     CrateMetadata,
     DEFAULT_KEY_COL_WIDTH,
@@ -59,11 +53,6 @@ use subxt::{
 };
 use subxt_signer::sr25519::Keypair;
 
-use std::{
-    option::Option,
-    path::Path,
-};
-
 pub use balance::{
     BalanceVariant,
     TokenMetadata,
@@ -73,6 +62,7 @@ pub use call::{
     CallExec,
     CallRequest,
 };
+pub use contract_artifacts::ContractArtifacts;
 pub use contract_info::{
     fetch_all_contracts,
     fetch_contract_info,
@@ -115,132 +105,6 @@ pub use upload::{
 pub type Client = OnlineClient<DefaultConfig>;
 pub type Balance = u128;
 pub type CodeHash = <DefaultConfig as Config>::Hash;
-
-/// Contract artifacts for use with extrinsic commands.
-#[derive(Debug)]
-pub struct ContractArtifacts {
-    /// The original artifact path
-    artifacts_path: PathBuf,
-    /// The expected path of the file containing the contract metadata.
-    metadata_path: PathBuf,
-    /// The deserialized contract metadata if the expected metadata file exists.
-    metadata: Option<ContractMetadata>,
-    /// The Wasm code of the contract if available.
-    pub code: Option<WasmCode>,
-}
-
-impl ContractArtifacts {
-    /// Load contract artifacts.
-    pub fn from_manifest_or_file(
-        manifest_path: Option<&PathBuf>,
-        file: Option<&PathBuf>,
-    ) -> Result<ContractArtifacts> {
-        let artifact_path = match (manifest_path, file) {
-            (manifest_path, None) => {
-                let crate_metadata = CrateMetadata::from_manifest_path(
-                    manifest_path,
-                    contract_build::Target::Wasm,
-                )?;
-
-                if crate_metadata.contract_bundle_path().exists() {
-                    crate_metadata.contract_bundle_path()
-                } else if crate_metadata.metadata_path().exists() {
-                    crate_metadata.metadata_path()
-                } else {
-                    anyhow::bail!(
-                        "Failed to find any contract artifacts in target directory. \n\
-                        Run `cargo contract build --release` to generate the artifacts."
-                    )
-                }
-            }
-            (None, Some(artifact_file)) => artifact_file.clone(),
-            (Some(_), Some(_)) => {
-                anyhow::bail!("conflicting options: --manifest-path and --file")
-            }
-        };
-        Self::from_artifact_path(artifact_path.as_path())
-    }
-    /// Given a contract artifact path, load the contract code and metadata where
-    /// possible.
-    fn from_artifact_path(path: &Path) -> Result<Self> {
-        tracing::debug!("Loading contracts artifacts from `{}`", path.display());
-        let (metadata_path, metadata, code) =
-            match path.extension().and_then(|ext| ext.to_str()) {
-                Some("contract") | Some("json") => {
-                    let metadata = ContractMetadata::load(path)?;
-                    let code = metadata.clone().source.wasm.map(|wasm| WasmCode(wasm.0));
-                    (PathBuf::from(path), Some(metadata), code)
-                }
-                Some("wasm") => {
-                    let file_name = path.file_stem()
-                        .context("WASM bundle file has unreadable name")?
-                        .to_str()
-                        .context("Error parsing filename string")?;
-                    let code = Some(WasmCode(std::fs::read(path)?));
-                    let dir = path.parent().map_or_else(PathBuf::new, PathBuf::from);
-                    let metadata_path = dir.join(format!("{file_name}.json"));
-                    if !metadata_path.exists() {
-                        (metadata_path, None, code)
-                    } else {
-                        let metadata = ContractMetadata::load(&metadata_path)?;
-                        (metadata_path, Some(metadata), code)
-                    }
-                }
-                Some(ext) => anyhow::bail!(
-                    "Invalid artifact extension {ext}, expected `.contract`, `.json` or `.wasm`"
-                ),
-                None => {
-                    anyhow::bail!(
-                        "Artifact path has no extension, expected `.contract`, `.json`, or `.wasm`"
-                    )
-                }
-            };
-
-        if let Some(contract_metadata) = metadata.as_ref() {
-            if let Err(e) = contract_metadata.check_ink_compatibility() {
-                eprintln!("{} {}", "warning:".yellow().bold(), e.to_string().bold());
-            }
-        }
-        Ok(Self {
-            artifacts_path: path.into(),
-            metadata_path,
-            metadata,
-            code,
-        })
-    }
-
-    /// Get the path of the artifact file used to load the artifacts.
-    pub fn artifact_path(&self) -> &Path {
-        self.artifacts_path.as_path()
-    }
-
-    /// Get contract metadata, if available.
-    ///
-    /// ## Errors
-    /// - No contract metadata could be found.
-    /// - Invalid contract metadata.
-    pub fn metadata(&self) -> Result<ContractMetadata> {
-        self.metadata.clone().ok_or_else(|| {
-            anyhow!(
-                "No contract metadata found. Expected file {}",
-                self.metadata_path.as_path().display()
-            )
-        })
-    }
-
-    /// Get the code hash from the contract metadata.
-    pub fn code_hash(&self) -> Result<[u8; 32]> {
-        let metadata = self.metadata()?;
-        Ok(metadata.source.hash.0)
-    }
-
-    /// Construct a [`ContractMessageTranscoder`] from contract metadata.
-    pub fn contract_transcoder(&self) -> Result<ContractMessageTranscoder> {
-        let metadata = self.metadata()?;
-        ContractMessageTranscoder::try_from(metadata)
-            .context("Failed to deserialize ink project metadata from contract metadata")
-    }
-}
 
 /// The Wasm code of a contract.
 #[derive(Debug)]
